fix: allow hash character in directory listings (#27)
Pulls in the new `gateway-lib` with the fix for allowing hash character
in directory listings.

Also updates all dependencies, which includes a change in dagula to
switch from our unixfs exporter fork back to `ipfs-unixfs-exporter`
(since the parallel block fetching from the fork was backported).

`ipfs-unixfs-exporter` now uses `.subarray()` everywhere (no buffer
copies) which is rad, but it also caused our tests to fail which was
tricky to track down the cause. I've commented heavily in the place
where the fix is.