feat: use rollup indexes (#46)
This PR allows Freeway to use rollup indexes. Rollups are multi-index
indexes (see https://github.com/alanshaw/cardex#multi-index-index) that
specify which blocks are in which CAR file(s) for a given DAG.

Freeway checks DUDEWHERE for a rollup index at `bafyROOT/.rollup.idx`.
If it finds one, it'll use it (and no other indexes) to obtain index
informations for all the blocks.