feat: use blockly indexes (#51)
This PR adds support for "block link indexes" as a fallback, if content
cannot be served by DUDEWHERE and SATNAV.

It binds a new bucket `BLOCKLY` (BLOCK+LYnks) and adds a new index
implementation that uses the files stored in it.

Blockly is a collection of
[multi-indexes](https://github.com/web3-storage/specs/blob/73c386b999cf30fb648987ff9dce0516c1d91137/CARv2%20MultiIndex.md)
for individual blocks. Each index contains the CID of a CAR file the
block may be found in and the block's byte offset. The index also
contains the _same detail for any links_ the block has.