feat: serve CAR blocks (#68)
This PR adds a new handler that allows freeway to serve CAR files
directly from R2 (CARPARK).

Requesting a CID from the gateway with the CAR codec `0x0202` will
invoke this new handler.

Obviously CARs are not served in the same way as regular IPFS data,
since we can serve the data directly from an R2 bucket. To deal with
this difference, there's a new middlewares that detects the CAR codec
and calls the correct handler, passing through to the regular handlers
if the CAR codec is not detected.

The handler supports `HEAD` requests, allowing you to get the file size
of the item you're requesting. It also supports the HTTP `Range` header,
allowing you to extract byte ranges from within a CAR, which you might
do if you fetch the index for the CAR.

The idea is that we can use these handlers to build a system that runs
on content-claims. The flow might be something like:

1. Get [partition
claim](https://github.com/web3-storage/content-claims#partition-claim)
for a given content CID
2. Get [inclusion
claims](https://github.com/web3-storage/content-claims#inclusion-claim)
for each CAR `part` in the partition claim
3. Fetch each CARv2 index from the gateway for each `includes` CID
specified in the inclusion claims (You must first fetch partition claims
for each index CID to find out which CAR file they can be found in).
4. Use HTTP `Range` requests to export the whole DAG or a sub-DAG
directly from the `parts`, using the index data as a guide for which
blocks to extract.

Note: if exporting a full DAG you might want to just stop after (1) and
import all data in all CARs into IPFS.