feat: content claims reads by default with fallback for old index sources (#93)
Makes content claims used by default. If no index was found from content
claims, the handler is returned wrapped with the old
`withIndexerSources` and `withDagula` (old)

Alternatively, we could add a new middleware and keep there the old
ones. We could inspect in these middlewares if we have a dagula instance
already, but I don't think that is a good way to go as we in the future
can just drop this fallback