refactor: Type check and restructure tests (#119)
This is more what I was getting at with
https://github.com/storacha/freeway/pull/116/files#r1787752932. This
avoids some of the stashing and mutation by putting more inside the `new
Promise()` construction itself.

I've also pulled `wranglerEnv` out into its own `const`, because it's
actually an _input_, not an output. Then I've made `workerInfo` a `let`
which is assigned when Wrangler boots, and is `undefined` before that,
signaling that Wrangler is not yet available. The previous code was
failing because there were `undefined` fields that aren't typed to allow
`undefined`. Which wasn't an issue until…

I also turned type checking on in `test/`, because apparently it wasn't
already. 😛 And lastly, I've ignored `.wrangler`, because Wrangler writes
to that when it runs.

---------

Co-authored-by: Felipe Forbeck <fforbeck@gmail.com>