refactor(handler): at opentelemtry instrumentation (#125)
# Goals

Gain insight into what's going on with freeway

# Implementation

When in doubt, instrument. This adds opentelemtry and a default
cloudflare open telemetry implementation, that at minimum tells you how
long fetch requests are taking. And it's.... very interesting...

![Screenshot 2024-11-01 at 1 35
23 AM](https://github.com/user-attachments/assets/98b558c7-d37b-4513-9ec3-3b24eee8b584)

This is a trace of about a 64mb file downloading. 

I dunno what to make of all it, except to say... ouch those CARPARK
downloads take a WHILE. I wonder if somehow accessing them through
public URLs is a lot slower?