feat(indexer): probabilistic feature flag (#136)
### Context

Implemented a new logic to determine whether the "indexing service"
should be enabled for a given request.

1. **Query Parameter Check**: The code checks if the request URL
contains a query parameter `ff` with the value `indexing-service`. If
this parameter is present, it indicates that the indexing service should
be enabled for this request.

2. **Ramp-Up Probability**: If the query parameter is absent, the code
uses a probabilistic approach to decide whether to enable the indexing
service. This is done using a "ramp-up probability," which represents
the likelihood of enabling the service. The probability is defined by
the environment variable `FF_RAMP_UP_PROBABILITY`. A default value of 0
(or 0%) is used if this variable is not set.

3. **Random Selection**: The code uses `Math.random()` to generate a
random number between 0 and 1. If this number is less than or equal to
the ramp-up percentage, the indexing service is enabled.

### Related Issues
- https://github.com/storacha/project-tracking/issues/223