fix: handle options req for cors preflight (#137)
When we use the Console App to create a space, it will send an OPTIONS,
and then a POST request to the UCAN Invocation Server, which is behind
the CORS handler, and that CORS Handler doesn't support OPTIONS requests
for CORS preflight. I've added a new handler to process that OPTIONS
request.