Intercept fetches to R2 and use direct CARPARK pull (#141)
# Goals

When we fetch blocks from an R2 CARPark URL, we appear to get TTFB
around 400ms. Interestingly, if we instead use Cloudflare Worker's
direct R2 access, our TTFB drops to around 200ms. Bandwidth doesn't seem
to change much, but this is still an important win.

# Implementation

- Using a seperate change in blob-fetcher to allow passing a custom
fetch implementation, we now write a version of fetch that checks URLs
to see if they are CAR Park URLs. if they are, instead of using the
fetch API, we use our direct connection to CARPark in the worker to
perform the fetch request.
- Pass this custom fetch to the batching fetch used by content claims
dagula
- Also, to assemble a browser like Response object use the existing code
in withCarBlockHandler, factored out to utility file
- In the process of doing this, I discovered a bug in that code where it
was treating the last byte in the range as exclusive to the returned
content rather inclusive, and this was undetected as the test was
expecting the wrong result.

# For Discussion

Comparison traces, note the average TTFB of 200 vs 400 on requests to R2

Regular fetch:
![Screenshot 2025-01-03 at 8 42
11 PM](https://github.com/user-attachments/assets/78981630-8614-4060-8a3e-20e6b2868924)

With intercepted direct gets to R2:
![Screenshot 2025-01-04 at 3 48
55 PM](https://github.com/user-attachments/assets/52a3857a-f3eb-46e0-85c9-5337ec0491dc)