fix(service): properly handle attested delegations (#190)
- use the full signer + did:web as the authority for the server
- do not fail when multiple delegations are stored by content/serve
- pass blocks with each stored delegation when serving
- allow alternate authority on content serve delegations

# Why

- The first issue we're having is attestations must match the authority,
and we were using the signer, as opposed to the identity, so the dids
did not match
- Second, we didn't support storing multiple delegations, which is
problematic cause the `claim` check in the implementation of
access/delegate will fail if you don't also give it the attestation
- Third, we were not passing in attached blocks with the our delegations
- Finally, it's never come up before, but who exactly is the "authority"
on a storing of space/content/serve? Here, my belief is it should be the
upload-service, as it's the ultimate arbiter of whether egress can be
billed. I don't know if the upload service is properly checking each
submitted content serve for validity, but if it is, its authority will
be itself. So I added facilities to use the upload service as the
authority -- of note, the authority does NOT require a private key, so
don't worry we're just storing the upload service pub key in
wrangler.toml