feat: add a stripped down server for debugging (#196)
It's tricky to debug freeway in production as the deployed instance has
a lot of traffic and it's difficult to pinpoint requests and responses
and errors in the logs.

This PR adds a stripped down version of freeway that you can run locally
or deploy to cloudflare that fetches its data from the production
network.